+++ TimeIsMoney/finishResultViewController.m
@@ -55,7 +55,7 @@
     hours = app.prjTime/3600;
     minutes = (app.prjTime%3600)/60;
     seconds = (app.prjTime%3600)%60;
-    self.resultTimeLabel.text = [NSString stringWithFormat:@"%02ld:%02ld:%02ld",hours,minutes,seconds];
+    self.resultTimeLabel.text = [NSString stringWithFormat:@"%02ld:%02ld:%02ld",(long)hours,(long)minutes,(long)seconds];
     
     
     //resultCostLabelに報酬額から総コストを引いた金額を記入
@@ -63,7 +63,7 @@
     flt = app.prjTime/ichienByousu;
     cost = flt;
     cost = app.housyu - cost;
-    self.resultCostLabel.text = [NSString stringWithFormat:@"%ld",cost];
+    self.resultCostLabel.text = [NSString stringWithFormat:@"%ld",(long)cost];
     
     //resultJikyuLabelに報酬額をかかった時間で割った「時給」を記入
     resultJikyu = (app.housyu/app.prjTime)*3600;
@@ -71,7 +71,7 @@
         NSNumber *num = [NSNumber numberWithFloat:app.housyu]; //float型を編集
         self.resultJikyuLabel.text = [NSString stringWithFormat:@"%@",num];
     }else{
-        self.resultJikyuLabel.text = [NSString stringWithFormat:@"%ld",resultJikyu];
+        self.resultJikyuLabel.text = [NSString stringWithFormat:@"%ld",(long)resultJikyu];
     }
 }
